requirements.txt
# Minimal deps for the starter app
Pillow>=10.0.0
Pillow>=10.0.0
requests>=2.32.0

# Optional: enable tests
pytest>=8.0.0
